refactor patches to Jekyll module
