+++ lib/jekyll/data_path.rb
@@ -1,0 +1,7 @@
+module Jekyll
+  class Theme
+    def data_path
+      path_for "_data".freeze
+    end
+  end
+end
