--- lib/jekyll/unified_payload_drop.rb
@@ -1,29 +1,0 @@
-# encoding: UTF-8
-
-module Jekyll
-  module Drops
-    class UnifiedPayloadDrop < Drop
-      mutable true
-
-      attr_accessor :page, :layout, :content, :paginator
-      attr_accessor :highlighter_prefix, :highlighter_suffix
-
-      def jekyll
-        JekyllDrop.global
-      end
-
-      def site
-        @site_drop ||= JekyllData::ThemedSiteDrop.new(@obj)
-      end
-
-      def theme
-        @theme_drop ||= site[site.theme.name]
-      end
-
-      private
-      def fallback_data
-        @fallback_data ||= {}
-      end
-    end
-  end
-end
